refactor: move unlock dialog to avoid cycle dependencies (#378)
