+++ packages/app/src/systems/Core/components/index.tsx
@@ -9,3 +9,5 @@ export * from './RouteGuard';
 export * from './TestWrapper';
 export * from './VisibilityButton';
 export * from './RouterProvider';
+export * from './UnlockDialog';
+export * from './UnlockForm';
