+++ packages/app/src/systems/Core/hooks/index.tsx
@@ -1,1 +1,2 @@
 export * from './useIsLogged';
+export * from './useUnlockForm';
