+++ packages/app/src/systems/DApp/components/index.tsx
@@ -1,3 +1,1 @@
 export * from './ConnectInfo';
-export * from './UnlockDialog';
-export * from './UnlockForm';
