+++ packages/app/src/systems/DApp/hooks/index.tsx
@@ -1,2 +1,1 @@
 export * from './useSignatureRequest';
-export * from './useUnlockForm';
